requirements.txt
alembic==1.1.0
alipay==0.7.4
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
Babel==2.7.0
backcall==0.1.0
backports.ssl-match-hostname==3.7.0.1
bcrypt==3.1.7
beautifulsoup4==4.8.0
billiard==3.3.0.23
bs4==0.0.1
captcha==0.3
celery==3.1.26.post2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
decorator==4.4.0
Django==2.0.4
django-celery==3.3.1
django-core==1.4.1
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-formtools==2.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
dwebsocket==0.5.11
et-xmlfile==1.0.1
fake-useragent==0.1.11
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
flower==0.9.3
httplib2==0.13.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
kombu==3.0.37
lxml==4.4.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
mutagen==1.42.0
mysqlclient==1.4.4
openpyxl==3.0.3
paramiko==2.6.0
parsel==1.5.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
pipreqs==0.4.9
prompt-toolkit==2.0.9
psutil==5.6.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pycryptodome==3.9.0
pycryptodomex==3.7.2
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pymongo==3.9.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
PyQt5==5.13.2
PyQt5-sip==12.7.0
pyte==0.8.0
python-alipay-sdk==1.10.1
python-dateutil==2.8.0
python-editor==1.0.4
python-memcached==1.59
pytz==2019.2
pywin32==224
qiniu==7.2.6
queuelib==1.5.0
redis==3.3.8
requests==2.22.0
scapy==2.4.3
Scrapy==1.7.3
scrapy-redis==0.6.8
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.3
SQLAlchemy==1.3.8
sqlparse==0.3.0
termcolor==1.1.0
thefuck==3.29
tornado==5.1.1
traitlets==4.3.2
Twisted==19.7.0
urllib3==1.25.3
vine==1.3.0
virtualenv==16.7.5
virtualenvwrapper-win==1.2.5
vue==0.0.1
w3lib==1.21.0
wcwidth==0.1.7
Werkzeug==0.15.5
win-unicode-console==0.5
WTForms==2.2.1
xlrd==1.2.0
xlwt==1.3.0
yarg==0.1.9
zipp==0.6.0
zope.interface==4.6.0
